Move folder model initialization code to LoadoutGroupFilesProvider.cs, next to the file model creation
